Fix and composer.lock

+++ app/Traits/PasswordLessAuth.php
@@ -18,25 +18,25 @@ trait PasswordLessAuth
      *
      * @param $token
      *
+     * @return \Illuminate\Http\Response|\Symfony\Component\HttpFoundation\Response|void
      * @throws ValidationException
      *
-     * @return \Illuminate\Http\Response|\Symfony\Component\HttpFoundation\Response|void
      */
     public function passwordlessLogin(Request $request, $token)
     {
         if ($this->hasTooManyLoginAttempts($request)) {
             $this->fireLockoutEvent($request);
-
+            
             return $this->sendLockoutResponse($request);
         }
         if ($this->attemptPasswordlessLogin($token, $request)) {
             return $this->sendLoginResponse($request);
         }
         $this->incrementLoginAttempts($request);
-
+        
         return $this->sendFailedLoginResponse($request);
     }
-
+    
     /**
      * Attempt to log the user into the application.
      *
@@ -46,9 +46,15 @@ trait PasswordLessAuth
      */
     protected function attemptPasswordlessLogin($token, Request $request)
     {
-        $token = PasswordlessLoginToken::with('user')->whereToken($token)->first();
+        $token = PasswordlessLoginToken::with('user')->has('user')->whereToken($token)->first();
+        
+        if ( ! $token) {
+            return false;
+        }
+        
         if ($token->user instanceof User) {
             $token->delete();
+            
             return $this->guard()->login($token->user);
         }
